+++ CHANGELOG.md
@@ -5,7 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-None
+### Added
+- Srdimport: disable importing of monsters with unsupported features
+  - subtypes
+  - types: aberration, animal, humanoid and outsider
+
+### Fixed
+- Calculations: implement proper support for monster type class skills
+
+### Security
+- Update dependencies
 
 ## 0.1.0 - 2019-04-16
 ### Added